requirements.txt
alembic==1.14.0
amqp==5.3.1
aniso8601==9.0.1
anyio==4.7.0
asn1crypto==1.5.1
attrs==24.3.0
base58==2.1.1
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coincurve==20.0.0
colorama==0.4.6
cytoolz==1.0.1
eth-hash==0.7.0
eth-typing==5.0.1
eth-utils==5.1.0
eth_abi==5.1.0
flasgger==0.9.7.1
Flask==3.1.0
Flask-Admin==1.6.1
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
load-dotenv==0.1.0
Mako==1.3.8
MarkupSafe==3.0.2
mistune==3.0.2
packaging==24.2
parsimonious==0.10.0
prompt_toolkit==3.0.48
pycparser==2.22
pycryptodome==3.21.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rpds-py==0.22.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
toolz==1.0.0
tronpy==0.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
WTForms==3.1.2
